Update activeChoice after adding newField(from multiple-choice)
When pressing enter on empty choice new form-field should be added by deleting the empty choice. While adding new field  the active-choice state is persist that of deleted one.
Because of this while going back to the multiple-choice field. The choice is not focused properly.

This solves the focusing issue while navigating from other form fields into multiple-choice field.

Resolves: #136